Allow iOS dev builds to pair with release Macs
