+++ ios/Config/Shared.xcconfig
@@ -33,15 +33,13 @@ CMUX_DEV_TAG =
 // PresenceServiceConfiguration.resolvedServiceBaseURL.
 CMUX_PRESENCE_BASE_URL =
 
-// Auth-environment override, baked into the CMUXAuthEnvironment Info.plist key
-// so a normally-tapped dev device build (no shell env) can sign in against the
-// PRODUCTION Stack project and pair with a release (beta/stable) Mac — a
-// DEBUG build otherwise signs in to the development project, whose user ids
-// can never match a release Mac's QR account binding
-// (https://github.com/manaflow-ai/cmux/issues/7145). Empty by default
-// (DEBUG -> development, Release -> production). ios/scripts/reload.sh
-// --prod-auth sets it to "production". Keep the Info.plist key name in sync
-// with MobileAuthComposition.authEnvironmentInfoPlistKey.
+// Auth-environment override, baked into the CMUXAuthEnvironment Info.plist key.
+// Empty means the app uses its build default (DEBUG -> development, Release ->
+// production). ios/scripts/reload.sh now passes "production" by default so a
+// normally-tapped dev device build can sign in against the production Stack
+// project and pair with beta/stable/nightly Macs; --dev-auth passes
+// "development" for the local tagged-Mac dogfood flow. Keep the Info.plist key
+// name in sync with MobileAuthComposition.authEnvironmentInfoPlistKey.
 CMUX_IOS_AUTH_ENV =
 
 // ==========================================
