+++ ios/cmuxPackage/Sources/cmuxFeature/AuthCoordinatorIdentityProvider.swift
@@ -9,20 +9,10 @@ import CmuxMobileShellModel
 struct AuthCoordinatorIdentityProvider: MobileIdentityProviding {
     private let coordinator: AuthCoordinator
 
-    /// Whether the coordinator signs in against the development Stack project
-    /// (``MobileAuthComposition/authEnvironment``), so pairing can explain a
-    /// cross-environment account-binding mismatch truthfully.
-    let isDevelopmentAuthEnvironment: Bool
-
     /// Wrap an auth coordinator as an identity provider.
-    /// - Parameters:
-    ///   - coordinator: The coordinator owning the current session.
-    ///   - isDevelopmentAuthEnvironment: Whether the coordinator's resolved
-    ///     auth environment is development (its user ids belong to the dev
-    ///     Stack project).
-    init(coordinator: AuthCoordinator, isDevelopmentAuthEnvironment: Bool) {
+    /// - Parameter coordinator: The coordinator owning the current session.
+    init(coordinator: AuthCoordinator) {
         self.coordinator = coordinator
-        self.isDevelopmentAuthEnvironment = isDevelopmentAuthEnvironment
     }
 
     @MainActor var currentUserID: String? {
