+++ ios/scripts/reload.sh
@@ -12,21 +12,19 @@ Build, install, and launch the cmux iOS app with an isolated tag.
 By default this reloads only the simulator. Use --device to also reload the
 first available paired iPhone/iPad, or --device-only to skip the simulator.
 
-After install, the app is launched signed in (dogfood creds) and auto-paired to
-the tagged Mac app. Opt out granularly:
+By default, direct reloads use production auth and launch signed-out so they can
+pair with beta/stable/nightly Macs after you sign in with your real account.
+Pass --dev-auth to use dogfood credentials and auto-pair to the tagged Mac app.
+Opt out granularly:
   --no-sign-in   plain launch, no auto sign-in (implies no auto-pair)
   --no-attach    sign in, but do not auto-pair to the Mac
   --no-setup     plain install + launch (today's behavior)
 
-  --prod-auth    sign this DEV build in against PRODUCTION auth (bakes
-                 CMUXAuthEnvironment=production into Info.plist; the presence
-                 worker and API base follow the channel in-app), so it can
-                 pair with a real beta/stable Mac via QR. A plain dev build
-                 uses the development Stack project, whose user ids can never
-                 match a release Mac's QR account binding. Implies
-                 --no-sign-in (dogfood auto-login creds are dev-channel);
-                 sign in in-app with your real account and use the IN-APP
-                 scanner (the system Camera routes prod QRs to the beta app).
+  --prod-auth    deprecated no-op; production auth is now the default for DEV
+                 reloads so the build can pair with beta/stable/nightly Macs.
+  --dev-auth     build against the development Stack project. Use this for the
+                 local auto-sign-in/auto-pair dogfood flow only; it cannot pair
+                 with a production-auth Mac QR.
 
 Device signing uses the local Xcode account, or App Store Connect API
 credentials from ASC_API_KEY_ID, ASC_API_ISSUER_ID, ASC_API_KEY_PATH, or
@@ -61,14 +59,17 @@ RELOAD_SIMULATOR=1
 RELOAD_DEVICE=0
 ALLOW_PROVISIONING_UPDATES=1
 ALLOW_DEVICE_REGISTRATION=0
-# Auto-setup: after install + launch, sign in (inject dogfood creds) and auto-pair
-# to the tagged Mac app. Default ON; opt out granularly.
+# Auto-setup: under --dev-auth, after install + launch, sign in (inject dogfood
+# creds) and auto-pair to the tagged Mac app. Production-auth reloads skip this
+# because the dogfood credentials belong to the development Stack project.
+# Opt out granularly.
 NO_SIGN_IN=0
 NO_ATTACH=0
 NO_SETUP=0
-# --prod-auth: bake CMUXAuthEnvironment=production so the dev build signs in
-# against the production Stack project and can pair with a release Mac.
-PROD_AUTH=0
+# Auth channel baked into CMUXAuthEnvironment. DEV reloads default to production
+# so a sideloaded build can scan/pair with beta/stable/nightly Macs. --dev-auth
+# remains the explicit escape hatch for local dogfood credentials.
+AUTH_ENV_VALUE="production"
 
 while [[ $# -gt 0 ]]; do
   case "$1" in
@@ -133,7 +134,11 @@ while [[ $# -gt 0 ]]; do
       shift
       ;;
     --prod-auth)
-      PROD_AUTH=1
+      AUTH_ENV_VALUE="production"
+      shift
+      ;;
+    --dev-auth)
+      AUTH_ENV_VALUE="development"
       shift
       ;;
     -h|--help)
@@ -166,21 +171,18 @@ if [[ "$ALLOW_DEVICE_REGISTRATION" -eq 1 && "$ALLOW_PROVISIONING_UPDATES" -eq 0 
   exit 1
 fi
 
-# --prod-auth: point the build at the production auth channel so it can pair
-# with a real beta/stable Mac (https://github.com/manaflow-ai/cmux/issues/7145).
 # The value lands in the CMUXAuthEnvironment Info.plist key (a tapped device
 # build sees no shell env), read by MobileAuthComposition. Presence needs no
 # URL here: PresenceClient.resolvedServiceBaseURL follows the resolved auth
 # channel, so the worker URLs live only in Swift and cannot drift; an explicit
 # CMUX_PRESENCE_BASE_URL still wins as before.
-CMUX_IOS_AUTH_ENV_VALUE=""
-if [[ "$PROD_AUTH" -eq 1 ]]; then
-  CMUX_IOS_AUTH_ENV_VALUE="production"
+CMUX_IOS_AUTH_ENV_VALUE="$AUTH_ENV_VALUE"
+if [[ "$AUTH_ENV_VALUE" == "production" ]]; then
   # The dogfood auto-login creds are dev-Stack-project accounts; against
   # production auth they cannot sign in. Launch plain and sign in in-app with
   # the same account as the Mac you want to pair with.
   if [[ "$NO_SETUP" -eq 0 && "$NO_SIGN_IN" -eq 0 ]]; then
-    echo "==> --prod-auth: skipping auto sign-in/auto-pair (dogfood creds are dev-channel); sign in in the app"
+    echo "==> production auth: skipping auto sign-in/auto-pair (dogfood creds are dev-channel); sign in in the app"
     NO_SIGN_IN=1
   fi
 fi
@@ -558,11 +560,16 @@ Bundle id:
 Simulator:
   $SIMULATOR_NAME ($SIM_ID)
 EOF
-  if [[ "$PROD_AUTH" -eq 1 ]]; then
+  if [[ "$AUTH_ENV_VALUE" == "production" ]]; then
     cat <<EOF
 Auth environment:
-  production (--prod-auth): sign in in-app with your real account, then pair
+  production (default): sign in in-app with your real account, then pair
   with your beta/stable Mac using the in-app QR scanner.
+EOF
+  else
+    cat <<EOF
+Auth environment:
+  development (--dev-auth): local dogfood auto sign-in/auto-pair is enabled.
 EOF
   fi
 }
@@ -679,11 +686,16 @@ Bundle id:
 Device:
   $selected_device_name ($selected_device_id)
 EOF
-  if [[ "$PROD_AUTH" -eq 1 ]]; then
+  if [[ "$AUTH_ENV_VALUE" == "production" ]]; then
     cat <<EOF
 Auth environment:
-  production (--prod-auth): sign in in-app with your real account, then pair
+  production (default): sign in in-app with your real account, then pair
   with your beta/stable Mac using the in-app QR scanner.
+EOF
+  else
+    cat <<EOF
+Auth environment:
+  development (--dev-auth): local dogfood auto sign-in/auto-pair is enabled.
 EOF
   fi
 }
